Add support for passing fragment size
It's been about a decade, and links are faster, so it's more valuable to
be able to test combinations of parallelism and fragment size for
performance.

In particular, aria can be about 2x faster on links that yield about 2
gigabits/second over the Internet.  While it has the advantage of not
needing to pipeline, it should be possible to close the gap somewhat.
It uses a higher parallelism by default.